requirements.txt
flask
flask_cors
tenssorflow
numpy
pymongo
pillow